refactor: co-locate loader utilities with loaders
